+++ src/core/receiver/gnss_flowgraph.h
@@ -214,6 +214,7 @@ private:
     std::list<Gnss_Signal> available_GAL_1B_signals_;
     std::list<Gnss_Signal> available_GAL_5X_signals_;
     std::list<Gnss_Signal> available_GAL_7X_signals_;
+    std::list<Gnss_Signal> available_GAL_E6_signals_;
     std::list<Gnss_Signal> available_GLO_1G_signals_;
     std::list<Gnss_Signal> available_GLO_2G_signals_;
     std::list<Gnss_Signal> available_BDS_B1_signals_;
@@ -228,6 +229,7 @@ private:
         evGAL_1B,
         evGAL_5X,
         evGAL_7X,
+        evGAL_E6,
         evGLO_1G,
         evGLO_2G,
         evBDS_B1,